Halfway with the LineView

+++ Back-end/NEDa/src/main/java/com/TWNEDa/Earthquakes/EarthquakesController.java
@@ -23,8 +23,8 @@ public class EarthquakesController {
     }
 
     @RequestMapping(value="earthquakes/search",method = RequestMethod.GET)
-    public List<Earthquake> getFilteredEarthquakes(@Param(value="minFatalities") Integer minFatalities,
-                                                   @Param(value="maxFatalities") Integer maxFatalities,
+    public List<Earthquake> getFilteredEarthquakes(@Param(value="minFatalities") Long minFatalities,
+                                                   @Param(value="maxFatalities") Long maxFatalities,
                                                    @Param(value = "minMagnitude") Double minMagnitude,
                                                    @Param(value="maxMagnitude") Double maxMagnitude){
